apps/data-processing/requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
openpyxlpa
pyarrow
apscheduler

# Data processing
scipy>=1.10.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# API and web scraping
beautifulsoup4>=4.12.0
selenium>=4.15.0
aiohttp>=3.8.0
feedparser>=6.0.10

# Blockchain and crypto
web3>=6.0.0
starknet-py>=0.20.0
ccxt>=4.0.0

# Caching and message queues
redis>=4.6.0
celery>=5.3.0

# Async and concurrency
asyncio-throttle>=1.0.2
aiofiles>=23.2.1

# Data validation and serialization
pydantic>=2.4.0
marshmallow>=3.20.0

# HTTP clients and APIs
httpx>=0.25.0
urllib3>=2.0.0

# Time and scheduling
pytz>=2023.3
schedule>=1.2.0
croniter>=1.4.1
APScheduler>=3.10.4

# CLI and configuration
click>=8.1.0
configparser>=6.0.0
PyYAML>=6.0.1

# Utilities
loguru>=0.7.0
prometheus-client>=0.17.0
python-json-logger>=2.0.7

# Text processing and NLP
nltk>=3.8.1
textblob>=0.17.1
spacy>=3.7.0

# Financial calculations
QuantLib>=1.32
ta-lib>=0.4.28
yfinance>=0.2.22

# Image processing (for charts/graphs)
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Excel and CSV processing
openpyxl>=3.1.0
xlsxwriter>=3.1.9

# Monitoring and observability
sentry-sdk>=1.38.0
structlog>=23.2.0

# Security
cryptography>=41.0.0
PyJWT>=2.8.0
bcrypt>=4.0.1

# Environment and system
psutil>=5.9.0
python-multipart>=0.0.6

# Development and debugging
ipython>=8.15.0
rich>=13.6.0

# Optional: Machine Learning (uncomment if needed)
# tensorflow>=2.13.0
# torch>=2.0.0
# transformers>=4.35.0

# Optional: Additional databases (uncomment if needed)
# pymongo>=4.5.0
# elasticsearch>=8.10.0
# influxdb-client>=1.38.0